@@ -51,42 +49,27 @@
     "copy-webpack-plugin": "^13.0.0",
     "cross-env": "^7.0.3",
     "css-loader": "^7.1.2",
-    "dotenv": "^16.4.7",
-    "eslint": "^9.23.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-import-resolver-typescript": "^4.3.1",
-    "eslint-import-resolver-webpack": "^0.13.10",
+    "dotenv": "^16.5.0",
+    "eslint": "^9.28.0",
+    "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-n": "^15.0.0 || ^16.0.0",
-    "eslint-plugin-prettier": "^5.2.5",
-    "eslint-plugin-react": "^7.37.4",
-    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react": "^7.37.5",
     "favicons": "^7.2.0",
     "favicons-webpack-plugin": "^6.0.1",
-    "filemanager-webpack-plugin": "^8.0.0",
-    "fork-ts-checker-webpack-plugin": "^9.0.2",
+    "fork-ts-checker-webpack-plugin": "^9.1.0",
     "globby": "^14.1.0",
     "html-webpack-plugin": "^5.6.3",
     "mini-css-extract-plugin": "^2.9.2",
-    "postcss": "^8.5.3",
-    "postcss-loader": "^8.1.1",
-    "postcss-preset-env": "^10.1.5",
-    "postcss-scss": "^4.0.9",
     "prettier": "^3.5.3",
     "resolve-url-loader": "^5.0.0",
-    "sass": "^1.86.0",
-    "sass-loader": "^16.0.5",
-    "source-map-loader": "^5.0.0",
-    "style-loader": "^4.0.0",
     "terser-webpack-plugin": "^5.3.14",
     "ts-node": "^10.9.2",
     "tsconfig-paths-webpack-plugin": "^4.2.0",
-    "typescript": "^5.8.2",
-    "typescript-eslint": "^8.28.0",
-    "webpack": "^5.98.0",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.33.1",
+    "webpack": "^5.99.9",
     "webpack-cli": "^6.0.1",
     "webpack-ext-reloader": "^1.1.13",
-    "webpack-merge": "^6.0.1",
-    "wext-manifest-loader": "^3.0.0"
+    "webpack-merge": "^6.0.1"
   }
 }
